First proof of concept of advanced tiling
This feature works as follows:
When a window is already tiles to the given location a next monitor in this "direction" is searched to move the window there.
e.g. just tile twice "to the right" will first just tile it to the right, then tile to the right on the next monitor.